+++ src/components/form/form-field/form-field.test.js
@@ -64,7 +64,7 @@ describe("form-field", () => {
 					[{ type: "password" }, "form-input"],
 					[{ type: "textarea" }, "form-textarea"],
 					[{ type: "checkbox" }, "form-checkbox"],
-					[{ type: "radio-group", options: [] }, "radio-group"],
+					[{ type: "radio-group", options: [] }, "form-radio-group"],
 					[{ type: "button-group", options: [] }, "button-group"],
 				])("%s", ([props, component]) => {
 					const wrapper = mount({ props });