+++ src/components/form/form-input-group/form-input-group.test.js
@@ -1,11 +1,11 @@
 import { createMount } from "@unit/support/mount";
 import { describe, expect, test } from "vitest";
-import RadioGroup from "./radio-group.vue";
+import FormInputGroup from "./form-input-group.vue";
 
 const defaultProps = { options: ["pineapple", "banana", "coconut"] };
-const mount = createMount(RadioGroup, { props: defaultProps });
+const mount = createMount(FormInputGroup, { props: defaultProps });
 
-describe("radio-group", () => {
+describe("form-input-group", () => {
 	describe("Initialisation", () => {
 		test("should exist as a Vue component", () => {
 			const wrapper = mount();